# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/user/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "description": "UserCreate Param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserCreateResp"
                        }
                    }
                }
            }
        },
...
        "/user/list": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "user list",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "age",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "realname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.UserResp"
                            }
                        }
                    }
                }
            }
        },
        "/user/search": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "user search",
                "parameters": [
                    {
                        "description": "UserSearch Param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserSearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserResp"
                        }
                    }
                }
            }
        },
        "/user/{userId}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Fetch UserInfo profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserResp"
                        }
                    }
                }
            }
        },
        "/user/{userId}/delete": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "user delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModifyRowsResp"
                        }
                    }
                }
            }
        },
        "/user/{userId}/update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "userInfo update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UserUpdate Param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModifyRowsResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "condition.QueryCondition": {
            "type": "object",
            "required": [
                "comparison",
                "fieldname",
                "operator",
                "value"
            ],
            "properties": {
                "comparison": {
                    "type": "string",
                    "enum": [
                        "eq",
                        "ne",
                        "gt",
                        "ge",
                        "lt",
                        "le",
                        "in",
                        "like"
                    ]
                },
                "fieldname": {
                    "type": "string"
                },
                "operator": {
                    "type": "string",
                    "enum": [
                        "and",
                        "or"
                    ]
                },
                "value": {}
            }
        },
...
        "types.ModifyRowsResp": {
            "type": "object",
            "properties": {
                "rowsAffected": {
                    "type": "integer"
                }
            }
        },
        "types.UserCreateReq": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "realname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.UserCreateResp": {
            "type": "object",
            "properties": {
                "lastId": {
                    "type": "integer"
                }
            }
        },
        "types.UserResp": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "realname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.UserSearchReq": {
            "type": "object",
            "properties": {
                "query": {
                    "description": "pagination.Page",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/condition.QueryCondition"
                    }
                }
            }
        },
        "types.UserUpdateReq": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "realname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}